refactor: use existing module
